requirements.txt
appdirs==1.4.3
asgiref==3.3.1
black==20.8b1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-cleanup==5.1.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-reset-password==1.2.8
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
mypy-extensions==0.4.3
packaging==20.3
pathspec==0.8.1
pep517==0.8.2
Pillow==8.0.1
progress==1.5
PyJWT==2.0.0
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.5
regex==2020.11.13
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
